Update Comet integration to be optional, update code and tests
